refactor(code_gen): extract io builtins into builtins/io.rs
